+++ rollbar-wai/test/Rollbar/WaiSpec.hs
@@ -11,7 +11,7 @@ import qualified Network.Wai as W
 import qualified Network.Wai.Handler.Warp as W
 
 import Control.Concurrent (threadDelay)
-import Control.Monad (join, void)
+import Control.Monad (join)
 import Control.Monad.IO.Class
 import Data.Aeson
 import Data.IORef
@@ -62,15 +62,6 @@ spec = before getSettingsAndItemRef $
                 , requestUserIp = ""
                 }
             )
-
-    context "when the request handler throws" $
-      it "reports the exception type as the class" $
-        withApp $ \itemRef warpPort -> do
-          let url = http "localhost" /: "error"
-          void $ runReq
-            (defaultHttpConfig { httpConfigCheckResponse = \_ _ _ -> Nothing })
-            (req GET url NoReqBody bsResponse $ port warpPort)
-          threadDelay 500
           exception <- fmap (>>= itemException) $ readIORef itemRef
           fmap exceptionClass exception `shouldBe` Just "ErrorCall"
           fmap exceptionMessage exception `shouldBe` Just (Just "Boom")